+++ app/Services/Bonvoice/BonvoiceWebhookProcessorService.php
@@ -17,7 +17,6 @@ class BonvoiceWebhookProcessorService
         private readonly BonvoiceCallEventStore $callEventStore,
         private readonly BonvoiceLiveCallAssistService $liveCallAssistService,
         private readonly BonvoiceMissedCallRecoveryService $missedCallRecoveryService,
-        private readonly BonvoiceOutboundClickToCallLiveStatusService $outboundClickToCallLiveStatusService,
         private readonly BonvoiceIncomingCallLatency $incomingCallLatency,
     ) {}
 
@@ -62,7 +61,6 @@ class BonvoiceWebhookProcessorService
                 $this->liveCallAssistService->maybeNotify($callEvent);
                 $this->liveCallAssistService->maybeBroadcastAnsweredAutoOpen($callEvent, $previousStatus);
                 $this->liveCallAssistService->maybeBroadcastMissedPopupDismiss($callEvent, $previousStatus);
-                $this->outboundClickToCallLiveStatusService->maybeBroadcast($callEvent, $previousStatus);
             }
 
             if (! $options->suppressRecovery) {